+++ src/middleware.ts
@@ -214,6 +214,7 @@ async function handleRequest(context: APIContext, next: MiddlewareNext): Promise
     "/admin/categorias": "admin_siget_categorias",
     "/admin/ciclos": "admin_siget_ciclos",
     "/admin/secciones": "admin_siget_secciones",
+    "/admin/traslados": "admin_siget_traslados",
     "/admin/usuarios": "admin_siget_usuarios",
     "/admin/empresas": "admin_siget_empresas",
     "/base-de-conocimientos": "base_conocimientos",